requirements.txt
torch
numpy
scipy
matplotlib
pillow
tqdm
scikit-image